feat: add installer task
