+++ goreleaserInstallerTask/package.json
@@ -1,0 +1,9 @@
+{
+    "dependencies": {
+        "@types/node": "17.0.30",
+        "axios": "0.27.2",
+        "azure-pipelines-task-lib": "3.2.0",
+        "azure-pipelines-tool-lib": "1.3.0",
+        "semver": "7.3.7"
+    }
+}
