+++ goreleaserInstallerTask/task.json
@@ -1,0 +1,47 @@
+{
+  "id": "task_id",
+  "name": "goreleaser-installer",
+  "friendlyName": "GoReleaser Installer",
+  "description": "Install GoReleaser",
+  "category": "Utility",
+  "author": "Engin Diri",
+  "helpUrl": "https://goreleaser.com/intro/",
+  "helpMarkDown": "[Learn more about this task](https://goreleaser.com/intro/) ",
+  "releaseNotes": "First release",
+  "version": {
+    "Major": 0,
+    "Minor": 1,
+    "Patch": 0
+  },
+  "instanceNameFormat": "GoReleaser is a release automation tool for Go projects",
+  "inputs": [
+    {
+      "name": "version",
+      "type": "string",
+      "label": "Version of goreleaser to use",
+      "defaultValue": "latest",
+      "required": true,
+      "helpMarkDown": "Select the goreleaser version to use e.g. 'v0.183.0' (leave blank for latest)"
+    },
+    {
+      "name": "distribution",
+      "type": "pickList",
+      "label": "GoReleaser distribution",
+      "defaultValue": "goreleaser",
+      "required": true,
+      "options": {
+        "goreleaser": "GoReleaser",
+        "goreleaser-pro": "GoReleaser Pro"
+      },
+      "properties": {
+        "EditableOptions": "False"
+      },
+      "helpMarkDown": "GoReleaser distribution, either GoReleaser or GoReleaser Pro"
+    }
+  ],
+  "execution": {
+    "Node10": {
+      "target": "src//index.js"
+    }
+  }
+}
