+++ goreleaserInstallerTask/tsconfig.json
@@ -1,0 +1,25 @@
+{
+  "compilerOptions": {
+    "useUnknownInCatchVariables": false,
+    "target": "es6",
+    // revert to es6                      /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019' or 'ESNEXT'. */
+    "module": "commonjs"
+    /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */,
+    "outDir": "./dist"
+    /* Redirect output structure to the directory. */,
+    /* Strict Type-Checking Options */
+    "strict": true
+    /* Enable all strict type-checking options. */,
+    "noImplicitAny": false
+    /* Raise error on expressions and declarations with an implied 'any' type. */,
+    "skipLibCheck": true,
+    /* Module Resolution Options */
+    "esModuleInterop": true
+    /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */,
+    /* Source Map Options */
+    "inlineSourceMap": true
+    /* Emit a single file with source maps instead of having a separate file. */,
+    "inlineSources": true
+    /* Emit the source alongside the sourcemaps within a single file; requires '--inlineSourceMap' or '--sourceMap' to be set. */
+  }
+}
